--- core/tool_framework/utils/availability.py
@@ -1,104 +1,0 @@
-"""Shared tool-availability helpers for backend-aware integration sources.
-
-Tools that can delegate to a pre-injected fixture backend — as the
-synthetic harnesses under ``tests/synthetic/`` do — need an availability
-check that accepts either real connection-verified credentials or an
-injected ``_backend`` object.  Centralising those helpers here avoids
-cross-tool imports (e.g. ``DataDogMonitorsTool`` reaching into
-``DataDogLogsTool`` to borrow a helper) and keeps the pattern consistent
-across integrations.
-"""
-
-from __future__ import annotations
-
-
-def eks_available_or_backend(sources: dict[str, dict]) -> bool:
-    """Available when real EKS credentials are present OR a fixture backend is injected.
-
-    Used by EKS tool wrappers whose ``extract_params`` can delegate to a
-    mock ``eks_backend`` for synthetic tests.  Tools without backend
-    support continue to use the narrower check in
-    ``tools.eks_list_clusters_tool._eks_available``.
-
-    The ``_backend`` slot is reserved for fixture backends that implement
-    the EKS tool API (``list_pods``, ``get_pod_logs``, ...). Other backend
-    types that speak different protocols should be placed in their own
-    distinct source slots and are invisible to this check — the real EKS
-    tools stay deactivated for those modes.
-    """
-    eks = sources.get("eks", {})
-    return bool(eks.get("connection_verified") or eks.get("_backend"))
-
-
-def datadog_available_or_backend(sources: dict[str, dict]) -> bool:
-    """Available when real Datadog credentials are present OR a fixture backend is injected.
-
-    Used by Datadog tool wrappers whose ``extract_params`` can delegate
-    to a mock ``datadog_backend`` for synthetic tests.
-    """
-    dd = sources.get("datadog", {})
-    return bool(dd.get("connection_verified") or dd.get("_backend"))
-
-
-def groundcover_available_or_backend(sources: dict[str, dict]) -> bool:
-    """Available when real groundcover credentials are present OR a fixture backend is injected.
-
-    Used by groundcover tool wrappers whose ``extract_params`` can delegate to a
-    mock ``groundcover_backend`` for synthetic tests.
-    """
-    gc = sources.get("groundcover", {})
-    if gc.get("_backend"):
-        return True
-    return bool(gc.get("connection_verified") and gc.get("api_key"))
-
-
-def ec2_available_or_backend(sources: dict[str, dict]) -> bool:
-    """Available when real EC2/AWS credentials are present OR a fixture backend is injected.
-
-    Mirrors ``eks_available_or_backend``: gates EC2/ELB tool wrappers whose
-    ``extract_params`` can delegate to a mock ``aws_backend`` for synthetic tests.
-    The ``ec2`` source is available when resolved integrations or synthetic
-    backends provide EC2/ELB topology context.
-    """
-    ec2 = sources.get("ec2", {})
-    return bool(ec2.get("connection_verified") or ec2.get("_backend"))
-
-
-def cloudwatch_is_available(sources: dict[str, dict]) -> bool:
-    """Available when a CloudWatch source is present in the alert context.
-
-    CloudWatch uses IAM-based auth, so availability is gated on the source key
-    existing. Tool params like ``job_queue`` are alert-specific and provided by
-    the LLM.
-    """
-    return bool(sources.get("cloudwatch"))
-
-
-def signoz_available_or_backend(sources: dict[str, dict]) -> bool:
-    """Available when real SigNoz credentials are present OR a fixture backend is injected.
-
-    Used by SigNoz tool wrappers whose ``extract_params`` can delegate to a
-    mock ``signoz_backend`` for synthetic tests.
-    """
-    signoz = sources.get("signoz", {})
-    if signoz.get("_backend"):
-        return True
-    return bool(signoz.get("connection_verified") and signoz.get("url") and signoz.get("api_key"))
-
-
-def tempo_available_or_backend(sources: dict[str, dict]) -> bool:
-    """Available when a verified Tempo config is present OR a fixture backend is injected.
-
-    Used by the Tempo tool wrapper whose ``extract_params`` can delegate to a
-    mock ``tempo_backend`` for synthetic tests.
-    """
-    tempo = sources.get("tempo", {})
-    if tempo.get("_backend"):
-        return True
-    return bool(tempo.get("connection_verified") and tempo.get("url"))
-
-
-def hermes_available_or_backend(sources: dict[str, dict]) -> bool:
-    """Available when Hermes integration is connected or a fixture backend is injected."""
-    hermes = sources.get("hermes", {})
-    return bool(hermes.get("connection_verified") or hermes.get("_backend"))
