+++ integrations/aws/topology_helper.py
