+++ integrations/cloudwatch/tools/cloudwatch_batch_metrics_tool/__init__.py
@@ -6,9 +6,9 @@ from typing import Any
 
 from core.tool_framework.telemetry import report_run_error
 from core.tool_framework.tool_decorator import tool
-from core.tool_framework.utils.availability import cloudwatch_is_available
 from core.tool_framework.utils.compaction import truncate_list
 from integrations.aws.cloudwatch_client import get_metric_statistics
+from integrations.cloudwatch.availability import cloudwatch_is_available
 
 
 @tool(
