+++ integrations/ec2/tools/ec2_instances_by_tag_tool/__init__.py
@@ -11,12 +11,12 @@ import logging
 from typing import Any, cast
 
 from core.tool_framework.tool_decorator import tool
-from core.tool_framework.utils.availability import ec2_available_or_backend
-from core.tool_framework.utils.aws_topology_helper import (
+from integrations.aws.aws_sdk_client import execute_aws_sdk_call
+from integrations.aws.topology_helper import (
     build_ec2_summary,
     extract_ec2_instances_params,
 )
-from integrations.aws.aws_sdk_client import execute_aws_sdk_call
+from integrations.ec2.availability import ec2_available_or_backend
 
 logger = logging.getLogger(__name__)
