+++ integrations/groundcover/tools/__init__.py
@@ -7,14 +7,14 @@ from __future__ import annotations
 from typing import Any
 
 from core.tool_framework.tool_decorator import tool
-from core.tool_framework.utils.availability import groundcover_available_or_backend
-from core.tool_framework.utils.groundcover import (
+from integrations.groundcover.availability import groundcover_available_or_backend
+from integrations.groundcover.client import GroundcoverClient
+from integrations.groundcover.helpers import (
     DEFAULT_LOGS_QUERY,
     GCQL_GUIDANCE,
     base_extract_params,
     run_signal_query,
 )
-from integrations.groundcover.client import GroundcoverClient
 
 _LOGS_SOURCE = "groundcover_logs"
 _LOGS_MCP_TOOL = "query_logs"
@@ -185,7 +185,7 @@ def get_groundcover_query_reference(
 
 
 from core.tool_framework.tool_decorator import tool
-from core.tool_framework.utils.groundcover import (
+from integrations.groundcover.helpers import (
     DEFAULT_TRACES_QUERY,
     GCQL_GUIDANCE,
 )
