+++ tests/tools/test_elb_target_health_tool.py
@@ -6,7 +6,7 @@ from typing import Any
 
 import pytest
 
-from core.tool_framework.utils.aws_topology_helper import extract_target_health_params
+from integrations.aws.topology_helper import extract_target_health_params
 from integrations.elb.tools.elb_target_health_tool import _is_available, get_elb_target_health
