+++ tests/tools/test_github_file_contents_tool.py
@@ -43,9 +43,7 @@ def test_extract_params_maps_fields() -> None:
 
 
 def test_run_returns_unavailable_when_no_config() -> None:
-    with patch(
-        "core.tool_framework.utils.github_helpers.github_mcp_config_from_env", return_value=None
-    ):
+    with patch("integrations.github.helpers.github_mcp_config_from_env", return_value=None):
         result = get_github_file_contents(owner="org", repo="repo", path="README.md")
     assert result == {
         "source": "github",
@@ -66,11 +64,9 @@ def test_run_happy_path() -> None:
     }
     mock_config = MagicMock()
     with (
+        patch("integrations.github.helpers.github_mcp_config_from_env", return_value=None),
         patch(
-            "core.tool_framework.utils.github_helpers.github_mcp_config_from_env", return_value=None
-        ),
-        patch(
-            "core.tool_framework.utils.github_helpers.build_github_mcp_config",
+            "integrations.github.helpers.build_github_mcp_config",
             return_value=mock_config,
         ),
         patch(
