+++ tests/tools/test_github_helpers.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from unittest.mock import MagicMock, patch
 
-from core.tool_framework.utils.github_helpers import github_creds, resolve_github_mcp_config
+from integrations.github.helpers import github_creds, resolve_github_mcp_config
 from integrations.github_mcp import DEFAULT_GITHUB_MCP_MODE
 
 
@@ -51,7 +51,7 @@ def test_github_creds_omits_empty_defaults() -> None:
 def test_resolve_github_mcp_config_uses_env_when_no_overrides() -> None:
     env_config = MagicMock()
     with patch(
-        "core.tool_framework.utils.github_helpers.github_mcp_config_from_env",
+        "integrations.github.helpers.github_mcp_config_from_env",
         return_value=env_config,
     ):
         assert resolve_github_mcp_config(None, None, None) is env_config
@@ -62,12 +62,10 @@ def test_resolve_github_mcp_config_builds_when_token_present() -> None:
     built = MagicMock()
     with (
         patch(
-            "core.tool_framework.utils.github_helpers.github_mcp_config_from_env",
+            "integrations.github.helpers.github_mcp_config_from_env",
             return_value=env_config,
         ),
-        patch(
-            "core.tool_framework.utils.github_helpers.build_github_mcp_config", return_value=built
-        ) as builder,
+        patch("integrations.github.helpers.build_github_mcp_config", return_value=built) as builder,
     ):
         result = resolve_github_mcp_config(None, None, "ghp_test")
     assert result is built
@@ -80,7 +78,7 @@ def test_resolve_github_mcp_config_builds_when_token_present() -> None:
 def test_resolve_github_mcp_config_does_not_treat_default_mode_as_override() -> None:
     env_config = MagicMock()
     with patch(
-        "core.tool_framework.utils.github_helpers.github_mcp_config_from_env",
+        "integrations.github.helpers.github_mcp_config_from_env",
         return_value=env_config,
     ):
         assert resolve_github_mcp_config(None, DEFAULT_GITHUB_MCP_MODE, None) is env_config
