+++ tests/tools/test_github_issues_tool.py
@@ -32,9 +32,7 @@ def test_extract_params_maps_fields() -> None:
 
 
 def test_run_returns_unavailable_when_no_config() -> None:
-    with patch(
-        "core.tool_framework.utils.github_helpers.github_mcp_config_from_env", return_value=None
-    ):
+    with patch("integrations.github.helpers.github_mcp_config_from_env", return_value=None):
         result = search_github_issues(owner="org", repo="repo", query="crash")
     assert result == {
         "source": "github",
@@ -55,11 +53,9 @@ def test_run_happy_path() -> None:
     }
     mock_config = MagicMock()
     with (
+        patch("integrations.github.helpers.github_mcp_config_from_env", return_value=None),
         patch(
-            "core.tool_framework.utils.github_helpers.github_mcp_config_from_env", return_value=None
-        ),
-        patch(
-            "core.tool_framework.utils.github_helpers.build_github_mcp_config",
+            "integrations.github.helpers.build_github_mcp_config",
             return_value=mock_config,
         ),
         patch("integrations.github.tools.issues.call_github_mcp_tool", return_value=fake_result),
@@ -85,11 +81,9 @@ def test_run_tool_error() -> None:
     }
     mock_config = MagicMock()
     with (
+        patch("integrations.github.helpers.github_mcp_config_from_env", return_value=None),
         patch(
-            "core.tool_framework.utils.github_helpers.github_mcp_config_from_env", return_value=None
-        ),
-        patch(
-            "core.tool_framework.utils.github_helpers.build_github_mcp_config",
+            "integrations.github.helpers.build_github_mcp_config",
             return_value=mock_config,
         ),
         patch("integrations.github.tools.issues.call_github_mcp_tool", return_value=fake_result),
