+++ tests/tools/utils/test_eks_workload_helper.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import pytest
 
-from core.tool_framework.utils.eks_workload_helper import (
+from integrations.eks.workload_helper import (
     extract_cluster_params,
     extract_workload_params,
 )
