+++ tools/git_deploy_timeline_tool/__init__.py
@@ -24,7 +24,7 @@ from typing import Any
 
 from core.domain.types.incident_window import IncidentWindow
 from core.tool_framework.tool_decorator import tool
-from core.tool_framework.utils.github_helpers import (
+from integrations.github.helpers import (
     github_creds,
     github_source_available,
     normalize_github_tool_result,
